rag-wasde-groq/requirements.txt
lancedb
pymupdf
sentence-transformers
groq
python-dotenv
pandas
requests
